requirements.txt
backoff==2.2.1
duckdb==1.0.0
hydra-core==1.3.2
httpx==0.27.2
ipdb
ipykernel
ipython==8.23.0
matplotlib==3.8.0
numpy==1.23.5
omegaconf==2.3.0
openai==1.17.0
pandas==2.0.3
PyYAML==6.0.1
scipy
sympy==1.12
tqdm==4.65.0

